backend/requirements.txt
fastapi==0.115.0
PyJWT[crypto]>=2.10.1
uvicorn[standard]==0.30.0
pydantic==2.7.0
python-dotenv==1.0.1
httpx==0.27.0

# AI providers (uncomment according to your choice)
# anthropic==0.34.0
# openai==1.45.0

# Prompt Guard — meta-llama/Llama-Guard-4-12B via HuggingFace Inference API
# No local ML dependencies — HTTP call via httpx (already in deps).
# Requires HUGGINGFACE_TOKEN in .env (Meta license accepted on hf.co).
# If token is absent or API is unreachable, the guard disables itself automatically (fail-open).
